requirements.txt
customtkinter
SpeechRecognition
pyttsx3
python-docx
openpyxl
python-pptx
pyaudio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
